Move role module into sparate file
